requirements.txt
google-generativeai==0.8.3
requests==2.32.3
beautifulsoup4==4.12.3
markdownify==0.13.1
python-dotenv==1.0.1
